refactor: refactored course architecture
