+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { DatabaseModule } from '../database/database.module';
 import { UserModule } from '../modules/user/infra/modules/user.module';
 import { UserScoreModule } from '../modules/user-score/user-score.module';
-import { CourseModule } from '../modules/course/course.module';
+import { CourseModule } from '../modules/course/infra/modules/course.module';
 import { UserActivitiesAnsweredModule } from '../modules/user-activities-answered/user-activities-answered.module';
 import { UserCoursesConcludedModule } from '../modules/user-courses-concluded/user-courses-concluded.module';
 import { ActivityModule } from '../modules/activity/activity.module';
