+++ src/modules/activity/services/activity.service.spec.ts
@@ -5,13 +5,13 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Activity } from '../entity/activity.entity';
 import { UserCourseConcluded } from '../../user-courses-concluded/entity/user-courses-concluded.entity';
 import { User } from '../../user/entity/user.entity';
-import { CourseService } from '../../course/services/course.service';
-import { CourseRepository } from '../../course/repository/course.repository';
+import { CourseService } from '../../course/infra/services/course.service';
+import { CourseRepository } from '../../course/repositories/course.repository';
 import { ActivityRepository } from '../repository/activity.repository';
 import { UserCourseConcludedRepository } from '../../user-courses-concluded/repository/user-courses-concluded.repository';
 import { UserActivityAnsweredRepository } from '../../user-activities-answered/repository/user-activities-answered.repository';
 import { JWTProvider } from '../../user/providers/jwt.provider';
-import { CourseNotFoundException } from '../../course/domain/errors/CourseNotFound.exception';
+import { CourseNotFoundException } from '../../course/domain/dtos/errors/CourseNotFound.exception';
 import { UnprocessableDataException } from '../../../shared/domain/errors/UnprocessableData.exception';
 import { ActivityNotFoundException } from '../domain/errors/ActivityNotFound.exception';
 import { UserRepository } from '../../user/repositories/user.repository';
