+++ src/modules/course/domain/dtos/errors/CourseNotFound.exception.ts
