+++ src/modules/course/domain/dtos/requests/FindCourses.request.dto.ts
