+++ src/modules/course/domain/dtos/requests/FindMultipleCourses.request.dto.ts
