--- src/modules/course/dto/course.dto.spec.ts
@@ -1,7 +1,0 @@
-import { CourseDto } from './course.dto';
-
-describe('CourseDto', () => {
-  it('should be defined', () => {
-    expect(new CourseDto()).toBeDefined();
-  });
-});
