--- src/modules/course/dto/course.dto.ts
@@ -1,1 +1,0 @@
-export class CourseDto {}
