+++ src/modules/course/infra/db/entity/course.entity.spec.ts
