+++ src/modules/user/infra/modules/user.module.ts
@@ -11,7 +11,7 @@ import { UserController } from '../http/controllers/user.controller';
 import { DatabaseModule } from '../../../../database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Activity } from '../../../activity/entity/activity.entity';
-import { Course } from '../../../course/entity/course.entity';
+import { Course } from '../../../course/infra/db/entity/course.entity';
 import { UserCourseConcluded } from '../../../user-courses-concluded/entity/user-courses-concluded.entity';
 import { UserActivityAnswered } from '../../../user-activities-answered/entity/user-activities-answered.entity';
 import { UserScore } from '../../../user-score/entity/user-score.entity';
